pyproject.toml
[project]
name = "adk-agentic-logging"
version = "0.1.0"
description = "Wide Event Logging & Observability for Google ADK Agents"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
dependencies = [
    "opentelemetry-api>=1.15.0",
]

[project.optional-dependencies]
fastapi = ["fastapi", "uvicorn"]
flask = ["flask"]
django = ["django"]
dev = [
    "ruff", 
    "mypy", 
    "pytest", 
    "pytest-asyncio", 
    "httpx", 
    "starlette",
    "types-Flask",
    "django-stubs",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/adk_agentic_logging"]

[tool.ruff.lint]
select = ["E", "F", "I", "B"]
ignore = []

[tool.mypy]
strict = true
python_version = "3.10"
ignore_missing_imports = true
